requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
flask>=2.0.0
python-dotenv>=0.21.0
